Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>             V1          V2 cluster
#> 1   1.11399603  0.79001278       1
#> 2   1.23752261 -0.22889936       1
#> 3   0.59032509  0.73629359       1
#> 4   0.14891740  1.01893887       1
#> 5  -1.13560643 -1.56411201       1
#> 6   0.09531124  0.63124236       1
#> 7  -1.04811056 -0.56524811       1
#> 8  -0.73570197  1.22001245       1
#> 9   0.57369065 -0.09228869       1
#> 10  0.53819525 -1.20836894       1
#> 11  0.51249752 -1.39623270       1
#> 12 -0.45257462 -0.10700971       1
#> 13 -0.49818815 -1.20761131       1
#> 14 -0.44298833  0.77505156       1
#> 15  1.12249230 -1.22721247       1
#> 16  1.76598251 -0.11225842       1
#> 17 -0.67987924 -0.45587827       1
#> 18 -0.95266603 -0.36104764       1
#> 19 -0.10278256  1.11987385       1
#> 20  0.83792294 -0.16885927       1
#> 21 -0.15616783 -0.19971844       1
#> 22  1.83831269 -0.10985512       1
#> 23  0.43984747  0.97323791       1
#> 24 -1.45423035  0.27416108       1
#> 25  0.50084897 -0.63651821       1
#> 26  4.44902551  5.91369195       2
#> 27  5.07907367  6.09994543       2
#> 28  5.45545960  4.63189658       2
#> 29  7.66198040  4.92285768       2
#> 30  4.18922061  5.82066771       2
#> 31  5.90053565  5.77080684       2
#> 32  4.66700158  5.02056271       2
#> 33  5.63040943  4.98622441       2
#> 34  6.74805258  4.26696942       2
#> 35  4.11561902  3.84879071       2
#> 36  4.50437753  4.86009756       2
#> 37  3.68863299  3.96441407       2
#> 38  6.10956890  6.64604447       2
#> 39  5.05272700  3.71068471       2
#> 40  7.20664494  4.07509141       2
#> 41  4.24356818  6.19329820       2
#> 42  5.29415211  7.12076403       2
#> 43  6.02283051  6.67661655       2
#> 44  5.38887452  6.39436246       2
#> 45  6.34108596  4.53305633       2
#> 46  4.52955602  5.23110224       2
#> 47  4.32388054  7.34141265       2
#> 48  5.25591366  5.76537895       2
#> 49  4.91554841  5.47368827       2
#> 50  4.57415159  5.46151127       2