Getter method for the parsed peak table in a cpc object

Method that returns the characterized peak list with all calculated peak characteristics.

Takes a cpc object and a single character vector indicating a slot in the cpcParam object.

Takes a named list of parameters or another cpcParam object as argument. The element names in the list must correspond to a slot name in the cpcProcParam object.

Parses the peak table contained in the XCMSnExp object and stores it in the pt slot.

Prior to running this method the peak table must be parsed using the parsePeaklist or parseFeatures methods.

This is the main method run on a cpc object to characterize the peaks detected by XCMS. Prior to running this method the peak table must be parsed using the parsePeaklist method.

This method is run after processPeaks in order to remove the peaks that do not fulfill the criteria and creates a filtered XCMSnExp object and stores it in the xdFilt slot.

This method returns the unfiltered original XCMSnExp object.

This method returns the filtered XCMS object after running the CPC algorithm.

This method will return a vector containing the peak IDs of the peaks that were removed during the CPC filtering.

This method will return a vector containing the peak IDs of the peaks that were *not* removed during the CPC filtering.

Outputs a summary of the cpc object.

# S4 method for cpc
getPeaklist(x)

# S4 method for cpc
cpt(x)

# S4 method for cpc
getProcData(x, value = NULL)

# S4 method for cpc
setProcData(x) <- value

# S4 method for cpc
setProcData(x) <- value

# S4 method for cpc
getProcParams(x, value = NULL)

# S4 method for cpc
getParam(x, param)

# S4 method for cpc
setParam(x) <- value

# S4 method for cpc
parsePeaklist(x)

# S4 method for cpc
getChromatogram(x, id)

# S4 method for cpc
processPeaks(x)

# S4 method for cpc
getFilterOutcomes(x)

# S4 method for cpc
filterPeaks(x)

# S4 method for cpc
getOriginalXCMS(x)

# S4 method for cpc
getFilteredXCMS(x)

# S4 method for cpc
getRemovedPeaks(x)

# S4 method for cpc
getRetainedPeaks(x)

# S4 method for cpc
show(object)

Arguments

x

A cpc object

value

A named list of parameters.

id

The peak ID

object

A cpc object

Value

A data.frame of the peak table A data.frame of the characterized peak table Named list holding the parameter values A cpc object A cpc object A named list of process parameters. The parameter value. A cpc object. A cpc object A cpc_chrom object A cpc object A cpc object A XCMSnExp object A XCMSnExp object A vector of peak IDs A vector of peak IDs