Constructor for a cpcProcParam object

These objects are used to set the parameters for the peak processing. Filter criteria are used to filter out peaks that do not meet the criteria.

cpcProcParam(
  ppm = 50,
  min_pts = 7L,
  min_inf_width = 3,
  min_sn = 10,
  min_frac = 0.5,
  min_shoulder_pts = 3L,
  min_rounded_pts = 3L,
  min_intensity = NULL,
  interval_tf = NULL,
  min_w = 5L,
  max_w = 21L,
  smooth_method = "savgol",
  smooth_times = 2L,
  smooth_win = NULL,
  max_sigma = NULL,
  fit_emg = FALSE,
  sel_peaks = NULL,
  sel_files = NULL,
  verbose_output = FALSE,
  save_all = FALSE,
  plot = FALSE
)

Arguments

ppm

The ppm range for generating XICs

min_pts

Filter criteria for minimum points between peak boundaries

min_inf_width

Filter criteria for minimum points between peak inflection points. Should be set >=3 to avoid issues.

min_sn

Filter criteria for minimum signal-to-noise ratio.

min_frac

Filter criteria for minimum samples a peak is found in. This is currently not used as it is only applicable to feature filtering.

min_shoulder_pts

Minimum number of points between the inflection points of a shoulder peak (default: 3)

min_rounded_pts

Minimum number of points between the inflection points of a rounded peak (default: 3)

min_intensity

Filter criteria for minimum peak area.

interval_tf

A numeric vector of length 2 giving the min and max tailing factor for filtering of peaks (default: NULL means peaks will not be filtered based on this)

min_w

Minimum smoothing function window size (default: 5L)

max_w

Maximum smoothing function window size (default: 21L)

smooth_method

Smoothing method used during processing. Available: "savgol" for Savitzky-Golay smoothing or "mean" for moving mean smoothing.

smooth_times

Number of smooth iterations,

smooth_win

Width of the smoothing function. If set to NULL it will be determined from the peak sigma values in the XCMSnExp object.

max_sigma

Largest allowable sigmal value used to determine the peak widths from the XCMSnExp object.

fit_emg

logical indicating if EMG deconvolution should be performed. Note that this adds a significant amount of time to the processing but gives better estimates of peak characteristics.

sel_peaks

integer vector indicating a subset of peak indices to be processed from the XCMSnExp object. If NULL all peaks are processed.

sel_files

integer vector indicating a subset of file indices to be processed from the XCMSnExp object. If NULL all files are processed.

verbose_output

logical value indicating if output is to be given during processing.

save_all

logical indicating if all processing data should be saved. Caution: This will result in a very large object!

plot

logical indicating if the results should be plotted for each peak

Value

A cpcProcParam object