Getter method for the procData slot

Takes a named list with data as argument. If a data name already exists, it is updated with the new value and if it does not already exist, it is added.

This method will plot the processing result as a panel for review.

This method will perform the processing of a chromatogram on a cpc_chrom object. This method is called repeatedly by characterize_xcms_peaklist.

# S4 method for cpc_chrom
getProcData(x, value = NULL)

# S4 method for cpc_chrom
setProcData(x) <- value

# S4 method for cpc_chrom
plotPeak(x, plotEMG = TRUE, plotXCMS = TRUE, annotation = character(1))

# S4 method for cpc_chrom
processChromatogram(x)

Arguments

x

A cpc_chrom object.

value

A named list with data.

plotEMG

logical indicating if results from EMG deconvolution should be plotted (default: TRUE)

plotXCMS

logical indicating if XCMS reported peak bounds should be plotted (default: TRUE)

annotation

(Optional) A single string or a vector of strings for each plotted peak with additional annotations for the plot

Value

A named list with process data. A cpc_chrom object A cpc_chrom object.