Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>             V1          V2 cluster
#> 1  -0.05365551  0.61138545       1
#> 2   0.47769025 -1.13468941       1
#> 3   0.92215163  0.32418496       1
#> 4   1.54288800  0.97791921       1
#> 5  -0.56060251 -1.83323562       1
#> 6  -0.85492648  0.23047084       1
#> 7  -1.07206853  0.65537731       1
#> 8   0.54927974  0.37896550       1
#> 9   0.04392689 -0.15222370       1
#> 10 -2.17273702  0.72147509       1
#> 11 -1.01903378 -1.34828210       1
#> 12  0.55189959  0.45394288       1
#> 13  1.80569424  0.68253652       1
#> 14 -0.01340229 -1.99069472       1
#> 15 -0.51689820  0.85296157       1
#> 16 -0.50402097 -0.29473020       1
#> 17  0.70248330  0.57101818       1
#> 18  0.04016730 -1.55590693       1
#> 19  0.06340500  0.02039797       1
#> 20  1.86202090  0.59513179       1
#> 21 -1.12496299 -0.60067047       1
#> 22  0.28326399  1.69615421       1
#> 23  0.67111268  0.27005519       1
#> 24  1.17942659 -1.16820016       1
#> 25  0.63045079 -0.29032868       1
#> 26  6.46964624  5.84558509       2
#> 27  6.10568776  5.20635625       2
#> 28  4.19038333  6.73558255       2
#> 29  3.85400341  2.77798487       2
#> 30  5.15810808  6.10972503       2
#> 31  5.81508652  4.48733244       2
#> 32  5.93874088  2.96186510       2
#> 33  5.16939078  5.10279723       2
#> 34  6.65487094  5.22495818       2
#> 35  2.62783555  5.59674236       2
#> 36  5.25728648  4.54307483       2
#> 37  5.14288967  4.29824029       2
#> 38  5.26665841  3.95275106       2
#> 39  3.71795088  5.22227850       2
#> 40  3.75337002  5.48822591       2
#> 41  3.30829796  4.85738894       2
#> 42  4.04942139  3.44635194       2
#> 43  4.12390175  5.16721039       2
#> 44  5.63763854  5.28849176       2
#> 45  4.29532135  4.25434334       2
#> 46  5.88651605  4.43702703       2
#> 47  5.02013363  3.87688934       2
#> 48  4.20383691  5.19371473       2
#> 49  4.15796014  5.47294360       2
#> 50  5.52594994  5.46425801       2