Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>              V1          V2 cluster
#> 1   0.957533386 -1.72420107       1
#> 2   1.365638536 -0.65721061       1
#> 3   0.022561552 -1.84190143       1
#> 4  -0.305198077  1.08420954       1
#> 5  -0.404736936 -0.88994285       1
#> 6  -0.851209634 -1.40143080       1
#> 7  -0.302403304  1.78859112       1
#> 8   0.751327545  1.09732298       1
#> 9  -0.441861962 -1.06138043       1
#> 10 -0.510538687  2.70456606       1
#> 11  0.199663038  1.62645863       1
#> 12  0.157055216  0.51846962       1
#> 13  0.001443738 -0.75112330       1
#> 14  0.156278786  2.00024951       1
#> 15  0.495011020 -0.54560628       1
#> 16  1.869305764  0.19489256       1
#> 17 -1.115068983  1.48462229       1
#> 18  1.749031990 -0.08497431       1
#> 19  0.412081865 -0.64786340       1
#> 20 -0.073556712 -1.14905062       1
#> 21  1.278887443  0.21940848       1
#> 22  0.313491004  0.06807024       1
#> 23 -0.055961161  0.50451301       1
#> 24 -0.350546749 -0.48682101       1
#> 25  0.444220919  0.98825307       1
#> 26  5.244850551  4.29600147       2
#> 27  3.970373111  5.25211302       2
#> 28  6.042034916  4.76220035       2
#> 29  4.333479022  5.79491207       2
#> 30  5.279715120  5.69981334       2
#> 31  4.740810143  4.18026804       2
#> 32  5.519823453  3.04460324       2
#> 33  5.530032846  3.80193514       2
#> 34  4.160103482  5.91078530       2
#> 35  6.099418602  3.46873608       2
#> 36  5.364405888  7.52961688       2
#> 37  4.633577744  4.84937171       2
#> 38  4.387729359  6.06958353       2
#> 39  6.243491906  3.72679589       2
#> 40  5.599904530  5.89930744       2
#> 41  6.625567163  5.62235067       2
#> 42  5.674686393  5.05960624       2
#> 43  4.141751188  5.77169177       2
#> 44  4.592339119  5.62083794       2
#> 45  5.704765221  4.80204103       2
#> 46  4.512506716  6.35120770       2
#> 47  5.487827496  6.44890828       2
#> 48  5.197368115  4.90087991       2
#> 49  4.995892205  6.04437314       2
#> 50  4.677626352  4.31305996       2