Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>             V1          V2 cluster
#> 1   0.10453929 -0.18452737       1
#> 2  -0.94853099 -2.23988156       1
#> 3  -1.35544535 -1.04371974       1
#> 4  -1.87456605  0.66105717       1
#> 5   0.72109628 -0.52204811       1
#> 6   0.86781686  1.95133638       1
#> 7   0.91584428 -0.66080040       1
#> 8   0.57896654  1.43227142       1
#> 9   0.12362139 -1.53534031       1
#> 10 -0.80632829 -0.41865274       1
#> 11 -1.54049848  1.56965226       1
#> 12 -1.61082355 -0.46788736       1
#> 13 -0.43508623 -0.63684939       1
#> 14  0.41092551  1.11031104       1
#> 15 -0.25530203 -0.57067353       1
#> 16 -0.07135794  1.16165854       1
#> 17  1.18224341  0.75794761       1
#> 18  1.43732749 -0.04774466       1
#> 19  1.27585288 -1.92710163       1
#> 20 -1.98259470 -1.89636428       1
#> 21 -0.52260557  0.92179664       1
#> 22 -0.21552148 -0.02475218       1
#> 23  0.38391992  1.63830656       1
#> 24 -0.31922927  1.71680046       1
#> 25 -0.68842142  0.92161623       1
#> 26  3.70945259  5.92340724       2
#> 27  3.75776530  6.05311797       2
#> 28  3.56570146  3.73502049       2
#> 29  3.04598337  5.53690977       2
#> 30  7.01876282  6.74490659       2
#> 31  4.47271958  5.05825848       2
#> 32  4.85403611  4.80558498       2
#> 33  4.28406910  5.19353170       2
#> 34  4.76775637  4.31458200       2
#> 35  4.86819076  4.78008834       2
#> 36  5.76658044  5.17145222       2
#> 37  6.70617585  4.90275614       2
#> 38  2.76841221  6.21459622       2
#> 39  5.44668067  7.57963904       2
#> 40  5.60493697  4.65051309       2
#> 41  5.97222950  4.32425818       2
#> 42  4.13203884  3.52886776       2
#> 43  3.55125821  2.86114028       2
#> 44  5.30381440  5.43061483       2
#> 45  5.47564958  5.96132901       2
#> 46  5.90569137  5.58559110       2
#> 47  4.92401418  5.98881033       2
#> 48  4.06413861  6.19040578       2
#> 49  4.69765357  4.77169308       2
#> 50  4.48491529  5.61391415       2