Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>             V1          V2 cluster
#> 1   1.27576965  0.50590004       1
#> 2  -0.01341733  1.74822928       1
#> 3   0.53089556 -0.23130221       1
#> 4  -1.15198311  0.19950795       1
#> 5   1.60558503 -0.65298820       1
#> 6  -0.10264023 -0.23609520       1
#> 7   0.91453915  0.44450388       1
#> 8   0.15435489 -0.39915775       1
#> 9   1.22079278 -0.71288833       1
#> 10  0.17002046  1.18278154       1
#> 11  0.68007959 -1.45166528       1
#> 12  2.21650742  0.57255628       1
#> 13  0.09382264 -0.34450881       1
#> 14  1.29437209  0.03252966       1
#> 15  0.91763910 -1.67123392       1
#> 16 -0.74871601  1.72734720       1
#> 17  0.29895369 -0.38590619       1
#> 18 -0.43700814 -0.55546703       1
#> 19  0.89327058  0.06879076       1
#> 20  0.06119307 -1.17656447       1
#> 21  0.18050879  1.39544323       1
#> 22 -1.23398004 -0.28027327       1
#> 23  0.15137360 -1.80732826       1
#> 24  0.52800249 -0.03935897       1
#> 25  0.05136709 -1.22000903       1
#> 26  5.57501735  5.36975673       2
#> 27  5.90565163  4.63545955       2
#> 28  4.22662548  3.91133080       2
#> 29  6.68132072  5.38084000       2
#> 30  3.23416941  4.71795678       2
#> 31  5.00665789  4.96471757       2
#> 32  3.76583414  4.20182129       2
#> 33  5.78990856  4.70396646       2
#> 34  4.40837635  5.69845613       2
#> 35  5.00038559  7.15212355       2
#> 36  2.19625604  4.62204622       2
#> 37  5.32290458  5.11942165       2
#> 38  7.57295077  4.75868870       2
#> 39  5.55023213  4.62357092       2
#> 40  5.59469538  3.66106789       2
#> 41  5.66212839  5.70594152       2
#> 42  5.11293400  5.96411022       2
#> 43  4.46094643  4.65797103       2
#> 44  4.31910663  3.39067134       2
#> 45  4.29746915  5.14780917       2
#> 46  6.66037479  5.44876415       2
#> 47  5.11972951  4.61464834       2
#> 48  4.79931718  4.62334198       2
#> 49  3.64311686  4.29754409       2
#> 50  4.82985280  3.97648755       2