Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>             V1         V2 cluster
#> 1  -0.06807136 -0.9355602       1
#> 2  -0.02392718 -1.0948047       1
#> 3  -1.58183638 -0.2568178       1
#> 4  -1.50019472 -1.1437437       1
#> 5   1.14235149  0.3853810       1
#> 6   1.88956064 -2.2713426       1
#> 7   0.61799514 -1.7161641       1
#> 8   0.48960003 -1.3390299       1
#> 9   0.05677455  1.4916010       1
#> 10  1.78178166  0.6104530       1
#> 11 -1.45387156  0.1211288       1
#> 12  1.28611994 -1.1447884       1
#> 13 -0.77653917  0.5461884       1
#> 14 -0.08268066  0.3782938       1
#> 15 -1.73969756 -1.0358539       1
#> 16 -0.52805994 -0.7853500       1
#> 17 -0.23420421  0.7945453       1
#> 18  0.44703467  1.0506847       1
#> 19 -0.97715039  0.8510539       1
#> 20 -2.20101071  0.4381498       1
#> 21 -0.31916261 -0.5995586       1
#> 22 -0.05886678 -0.6569705       1
#> 23  0.74130380 -2.0012703       1
#> 24 -1.36297829  1.1745309       1
#> 25 -1.44345737 -0.2909761       1
#> 26  5.03808430  5.8268817       2
#> 27  4.41231871  5.0662953       2
#> 28  3.68732860  6.7921103       2
#> 29  6.79831437  5.2371736       2
#> 30  4.35642518  4.4352615       2
#> 31  5.66612410  3.2965474       2
#> 32  4.76007959  7.1773588       2
#> 33  5.71180889  3.5558917       2
#> 34  3.75318102  2.7903909       2
#> 35  5.03670465  3.9644913       2
#> 36  5.28789463  4.1068272       2
#> 37  3.47292045  3.6285940       2
#> 38  5.81111215  4.6573021       2
#> 39  6.26134139  4.8099052       2
#> 40  4.66540489  7.3971024       2
#> 41  4.05289370  5.4727261       2
#> 42  6.45163393  5.1545141       2
#> 43  5.48735324  7.3631734       2
#> 44  4.38549692  6.8826404       2
#> 45  5.19023335  3.7855268       2
#> 46  4.89560253  5.1141765       2
#> 47  4.98278235  5.3954288       2
#> 48  5.22940659  5.1926901       2
#> 49  5.46948762  5.1927780       2
#> 50  3.74774959  5.8376603       2