Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>             V1          V2 cluster
#> 1  -0.36245876  0.14343178       1
#> 2  -0.35072960 -1.99146955       1
#> 3  -0.35598609  0.85143045       1
#> 4  -0.45645895  1.03311723       1
#> 5  -1.87543182 -0.20425189       1
#> 6  -1.01126475  0.85733813       1
#> 7  -1.53448772 -0.59665495       1
#> 8  -0.02123953  0.34240354       1
#> 9   0.01843745 -0.55938284       1
#> 10 -0.76773782 -0.20565025       1
#> 11  0.49922965 -0.84480277       1
#> 12  0.28165354  1.75337821       1
#> 13  0.38638489  0.18777284       1
#> 14 -0.08654002  1.57077175       1
#> 15  0.79134176  0.48559883       1
#> 16 -0.42451810  0.23598126       1
#> 17 -0.07267263 -1.73892550       1
#> 18  0.84494455 -0.31808154       1
#> 19  0.49951114 -0.31204093       1
#> 20  0.28226524 -0.76334169       1
#> 21  0.26109944  0.42913180       1
#> 22 -1.66820228 -0.07448808       1
#> 23 -0.52024997 -0.70148987       1
#> 24  0.77524007  1.68781992       1
#> 25 -0.56862368  0.89273808       1
#> 26  6.25663909  6.06002032       2
#> 27  5.55512766  4.97712849       2
#> 28  5.64471396  4.41831815       2
#> 29  2.62987783  5.44196242       2
#> 30  5.37259598  4.84886585       2
#> 31  5.24328300  5.14641281       2
#> 32  6.42660827  5.69665448       2
#> 33  3.76229096  5.81217157       2
#> 34  5.44319715  4.47519045       2
#> 35  4.95376381  3.70810691       2
#> 36  6.37906583  4.63324351       2
#> 37  5.86056947  4.18346474       2
#> 38  6.37015300  4.82584857       2
#> 39  3.59421220  6.19313632       2
#> 40  5.32358059  5.94299647       2
#> 41  4.37851223  4.37342752       2
#> 42  4.67778092  7.55713520       2
#> 43  5.87344449  4.60935474       2
#> 44  5.36762805  3.16113186       2
#> 45  6.64427922  7.03244027       2
#> 46  5.53764809  4.92619177       2
#> 47  5.12329591  4.28408695       2
#> 48  5.28253502  4.13682888       2
#> 49  4.43727914  5.27755426       2
#> 50  5.95069135  5.41206855       2