Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>             V1          V2 cluster
#> 1  -1.18759852 -0.13152061       1
#> 2  -1.94405013 -0.76069729       1
#> 3  -0.47583505 -0.20097255       1
#> 4  -0.81355504  1.58716427       1
#> 5   2.56703903  0.86913623       1
#> 6  -0.38951621  0.37629953       1
#> 7  -0.83234373  0.53494717       1
#> 8  -0.13677355  1.75053917       1
#> 9   0.95598354  0.52537545       1
#> 10 -0.09785501 -0.45141754       1
#> 11  1.07052522 -0.65863429       1
#> 12  1.88506483  0.56121434       1
#> 13 -0.99978090  2.19995121       1
#> 14  0.14641781  0.45437453       1
#> 15  0.94564669 -0.02191928       1
#> 16 -1.02475333 -0.27937094       1
#> 17 -1.70431827  0.32788145       1
#> 18  1.08747013  0.97745001       1
#> 19  0.16375110 -0.96030244       1
#> 20 -0.54439732 -0.53188035       1
#> 21 -0.39604144 -0.04755185       1
#> 22  0.46943919  1.66448242       1
#> 23 -0.51919203  0.58653512       1
#> 24  0.39648454 -0.12368023       1
#> 25 -0.66440479 -0.52894155       1
#> 26  7.20846358  6.11799827       2
#> 27  5.27348137  4.65778788       2
#> 28  6.07024722  3.60985020       2
#> 29  5.44675996  6.61336844       2
#> 30  7.60474375  5.33998503       2
#> 31  3.69379474  4.41962576       2
#> 32  5.42650706  4.72541926       2
#> 33  3.68607956  5.87911280       2
#> 34  3.49045114  4.73907089       2
#> 35  5.60355046  6.23790907       2
#> 36  5.83733372  5.04922119       2
#> 37  6.26825794  6.53717733       2
#> 38  4.78631977  5.76847678       2
#> 39  3.69382666  7.57000291       2
#> 40  6.59921444  4.23288848       2
#> 41  4.32692926  5.21249442       2
#> 42  4.79715507  6.42509365       2
#> 43  5.42553180  7.25567158       2
#> 44  5.15595824  3.57839277       2
#> 45  4.41042744  4.68622411       2
#> 46  5.05533905  4.42413531       2
#> 47  4.93674282  5.30194321       2
#> 48  5.73825736  5.83796077       2
#> 49  5.33070698  5.15233749       2
#> 50  5.34995113  5.63988251       2