Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>             V1           V2 cluster
#> 1  -0.77108311 -0.798742765       1
#> 2  -0.25852256 -0.001038566       1
#> 3  -0.37607475 -1.083805559       1
#> 4  -1.89309145 -0.313252252       1
#> 5   1.03710730 -0.898807147       1
#> 6  -1.58511091  0.010994303       1
#> 7  -0.54645015 -1.034243789       1
#> 8  -0.30747904 -1.324499705       1
#> 9   1.26726243 -0.328608670       1
#> 10 -2.11880385 -0.389749637       1
#> 11  2.04491437  0.153648217       1
#> 12 -0.32353388  0.810696002       1
#> 13  0.12708820  0.382933877       1
#> 14  0.01379621 -1.714850266       1
#> 15  0.48524556  1.412517997       1
#> 16 -0.98551000 -2.037083411       1
#> 17 -0.40835099  2.746063610       1
#> 18  0.12414030 -0.829107948       1
#> 19  2.04859066 -1.459252037       1
#> 20  0.63508624  0.633950121       1
#> 21 -1.41020386  0.261161544       1
#> 22  1.66463542  0.195100197       1
#> 23 -0.69064237 -1.442889537       1
#> 24  0.18995711  0.980251846       1
#> 25 -0.87989183 -0.593962827       1
#> 26  5.45476285  5.586856744       2
#> 27  5.04359127  3.426399101       2
#> 28  2.52612238  5.590093532       2
#> 29  3.90133674  5.336443421       2
#> 30  5.81931614  4.923956015       2
#> 31  6.70993284  4.934077628       2
#> 32  4.14868970  5.077067768       2
#> 33  3.93892083  3.510819929       2
#> 34  6.13965634  5.523293909       2
#> 35  4.09856498  4.672798317       2
#> 36  6.00894185  6.504725020       2
#> 37  2.66144160  5.668319223       2
#> 38  5.53806984  4.580598111       2
#> 39  4.45097134  4.821560978       2
#> 40  5.99502739  5.659360668       2
#> 41  4.41795212  5.698853182       2
#> 42  3.58714264  5.360411023       2
#> 43  5.24308276  5.120662942       2
#> 44  5.13053164  6.648409000       2
#> 45  6.03405754  3.730395119       2
#> 46  4.96731479  4.015600081       2
#> 47  5.87905159  4.981314927       2
#> 48  5.44745126  5.362807371       2
#> 49  4.70327549  4.653696148       2
#> 50  5.22192789  5.009838757       2