Data Manipulation for CPC Calculation

Converts numeric matrix to data frame with necessary format for "manual" CPC() calculation.

Usage

CPCdata.frame(data, cols, clusters)

Arguments

data

a numeric n x k matrix or data frame.

cols

columns in data to be used for calculating CPC().

clusters

column in data giving cluster membership.

Value

Returns a data frame with dimensions identical to those of data.

Examples

data <- matrix(c(rnorm(50, 0, 1), rnorm(50, 5, 1)), ncol = 2, byrow = TRUE)
clusters <- matrix(c(rep(1, 25), rep(2, 25)), ncol = 1)
data <- cbind(data, clusters)
CPCdata.frame(data, 1:2, 3)
#>             V1            V2 cluster
#> 1  -1.42799410  0.4200093632       1
#> 2   0.04087738 -0.0001277202       1
#> 3   0.57099967  0.3270754069       1
#> 4   1.35622231  0.0949506585       1
#> 5  -2.19835237  0.7589893020       1
#> 6  -0.14690270  1.3629470757       1
#> 7  -1.57955603  0.2918387625       1
#> 8   0.70526661  0.1627965373       1
#> 9  -0.56650584 -1.6011670114       1
#> 10 -1.09216175 -2.1268953798       1
#> 11 -0.31234997 -1.8598629936       1
#> 12 -0.04056080 -1.5932893101       1
#> 13 -0.81780516 -0.1437941494       1
#> 14  0.60685273  0.3880564462       1
#> 15 -0.26560832 -1.2223285671       1
#> 16  1.78396099  0.7679942514       1
#> 17 -0.25858320  0.5173789094       1
#> 18 -0.08035905  0.1495805149       1
#> 19  1.07623880 -1.9016883455       1
#> 20  0.39956195  1.2139284743       1
#> 21  0.11891785 -0.7782673504       1
#> 22  1.01441083  0.5199891171       1
#> 23  0.80478026  0.4613313080       1
#> 24  0.12758441 -2.3550247310       1
#> 25 -1.58683691  0.1620380255       1
#> 26  5.49994356  6.7738815848       2
#> 27  5.92610100  5.3988544085       2
#> 28  5.24334680  7.1713624805       2
#> 29  4.33334559  4.8403640938       2
#> 30  3.94392181  5.6104679878       2
#> 31  1.44065167  6.4432637351       2
#> 32  6.17214063  5.8102318680       2
#> 33  7.41097990  5.4627878329       2
#> 34  4.68791005  4.1236754910       2
#> 35  6.47517384  4.6487121267       2
#> 36  4.24947574  4.0795114939       2
#> 37  4.78232367  4.2338907789       2
#> 38  5.59094985  4.4791673160       2
#> 39  4.43221274  6.5011139802       2
#> 40  4.81762879  6.8808211718       2
#> 41  3.81163051  3.6324676611       2
#> 42  4.28610725  4.6699447569       2
#> 43  4.50900542  4.6157972465       2
#> 44  5.00309968  3.6811085443       2
#> 45  5.19390785  4.7140370849       2
#> 46  3.78741205  5.1477365018       2
#> 47  5.08947172  4.8929944359       2
#> 48  5.53299998  5.3906148741       2
#> 49  5.40009849  2.6780243834       2
#> 50  5.61068584  3.4868779391       2